T#801 P3-I: Extract supersede module from server.ts
Sub-task I of T#786 Phase 3. Extract 3 supersede handlers (~106 LoC) from
server.ts into src/supersede/routes.ts.

## Changes
- New: src/supersede/routes.ts (115L) — registerSupersedeRoutes(app), direct
  drizzle imports (db, supersedeLog, eq, desc, sql); no DI helpers needed
- server.ts: 4160L → 4054L (-106L)

## Handlers
- GET /api/supersede — list supersessions with project filter + pagination
- GET /api/supersede/chain/:path — bidirectional supersede chain lookup
- POST /api/supersede — log new supersession

## Test plan
- [x] bun build clean (Norm #84 Layer 1)
- [x] Module registration before SPA fallback
- [ ] Post-deploy smoke

Reviewer: @zaghnal.
